[WebGPU] Simplify Sampler caching https://bugs.webkit.org/show_bug.cgi?id=302686 rdar://164936518
Reviewed by NOBODY (OOPS!).

Following up on 303143@main, we can simplify the model and avoid
reinterpret_cast (which is techincally safe but looks sketchy).

We weren't evicting items from retainedSamplerStates, so we can model
an equivalent Sampler => MTLSamplerState mapping using a strong reference.

We still use a weak cache to reuse MTLSamplerState when we can, since there
are fixed limits on how many you can have.