Introduce ThreadGroupSnapshot to encapsulate a list of threads that won't exit https://bugs.webkit.org/show_bug.cgi?id=302722 rdar://164974979
Reviewed by NOBODY (OOPS!).

This is motivated by the SaferCPP bot complaining about raw back-pointers from
Thread to ThreadGroup.

My goal is to eliminate those raw pointers, and also reduce overall complexity,
since this code has been subject to mysterious deadlocks in the past, and its
current state is the result of a speculative fix.

In theory, once we encapsulate a set of threads that won't exit, we don't need
an intricate handshake between Thread and ThreadGroup when it's time to exit.

In theory.